Move PasswordGenerator.ts to the common/function folder. Adjust references.
