+++ src/__tests__/box/modules/boxCommon.ts
@@ -23,7 +23,7 @@ import { JwtModule } from '@nestjs/jwt';
 import BoxAuthHandler from '../../../box/auth/BoxAuthHandler';
 import { DailyTaskService } from '../../../box/dailyTask/dailyTask.service';
 import { GroupAdminGuard } from '../../../box/auth/decorator/groupAdmin.guard';
-import { PasswordGenerator } from '../../../box/tester/passwordGenerator';
+import { PasswordGenerator } from '../../../common/function/passwordGenerator';
 import { TesterService } from '../../../box/tester/tester.service';
 import SessionStarterService from '../../../box/sessionStarter/sessionStarter.service';
 import { DailyTasksModule } from '../../../dailyTasks/dailyTasks.module';
