+++ src/box/box.module.ts
@@ -23,7 +23,7 @@ import BoxAuthHandler from './auth/BoxAuthHandler';
 import { DailyTaskController } from './dailyTask/dailyTask.controller';
 import { DailyTaskService } from './dailyTask/dailyTask.service';
 import { GroupAdminGuard } from './auth/decorator/groupAdmin.guard';
-import { PasswordGenerator } from './tester/passwordGenerator';
+import { PasswordGenerator } from '../common/function/passwordGenerator';
 import { TesterService } from './tester/tester.service';
 import { TesterController } from './tester/tester.controller';
 import SessionStarterService from './sessionStarter/sessionStarter.service';
