+++ src/common/function/passwordGenerator.ts
